[CP-3213] Refactored presenters layer by moving `serializeRecord` to base abstract class
