+++ libs/core/data-sync/presenters/base-presenter.ts
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export abstract class BasePresenter{
+  protected serializeRecord<Type>(values: string[][], columns: string[]): Type[] {
+    return values.map((item) => {
+      return columns.reduce((acc: Record<string, string>, value, index) => {
+        acc[value.trim()] = String(item[index]).trim()
+        return acc
+      }, {})
+    }) as unknown as Type[]
+  }
+}
